"Convert success parameter to boolean in PaymentResponse component and update router props"
This diff includes changes to the PaymentResponse.vue component and the src/router.js file. In the PaymentResponse.vue component, the success parameter is converted from a string to a boolean, and a console log is added to display the data type of success. Additionally, based on the value of success, the component will redirect to either the home or checkout page after 5 seconds. In the src/router.js file, the props option is added to the payment-response route to enable passing props through the route.