"Convert success param to boolean and add redirection logic"
This commit converts the success parameter from a string to a boolean value and adds redirection logic based on the value of the success parameter. If success is true, the user is redirected to the home page after 5 seconds, otherwise, the user is redirected to the checkout page after 5 seconds.